pyproject.toml
[project]
name = "finsightai"
version = "0.1.0"
description = "Conversational financial insights agent."
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "beautifulsoup4>=4.13.4",
    "bs4>=0.0.2",
    "chromadb>=1.0.5",
    "crawl4ai>=0.5.0.post8",
    "langchain-text-splitters>=0.3.8",
    "langsmith>=0.3.32",
    "logfire>=3.14.0",
    "markdown>=3.8",
    "nltk>=3.9.1",
    "notebook>=7.4.0",
    "ollama>=0.4.8",
    "openai>=1.75.0",
    "playwright>=1.51.0",
    "pydantic>=2.11.3",
    "pydantic-ai[logfire]>=0.1.3",
    "pytest>=8.3.5",
    "python-dotenv>=1.1.0",
    "requests>=2.32.3",
    "rich>=14.0.0",
    "ruff>=0.11.6",
    "sentence-transformers>=4.1.0",
    "streamlit>=1.44.1",
    "tiktoken>=0.9.0",
    "typer>=0.15.2",
    "weasyprint>=65.1",
    "weaviate-client>=4.13.2",
]
